+++ src/main/kotlin/view/MainContent.kt
@@ -13,30 +13,39 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.layout.wrapContentSize
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Card
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Snackbar
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.FrameWindowScope
+import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.WindowScope
+import androidx.compose.ui.window.WindowState
 import model.entity.Message
 import org.koin.core.parameter.parametersOf
 import org.koin.java.KoinJavaComponent.inject
-import view.extention.onInitialize
+import resource.Images
+import view.components.pages.DevicesPage
+import view.components.pages.Page
 import view.page.DevicePage
-import view.page.Page
 import view.page.SettingPage
 import viewmodel.DevicePageViewModel
 import viewmodel.DevicesPageViewModel
 import viewmodel.MainContentViewModel
 import viewmodel.SettingPageViewModel
+import viewmodel.ViewModel
 
 @Composable
 fun MainContent(windowScope: WindowScope, mainContentViewModel: MainContentViewModel) {
@@ -69,13 +78,12 @@ private fun MainPages(windowScope: WindowScope, mainViewModel: MainContentViewMo
                     val viewModel by inject<DevicesPageViewModel>(clazz = DevicesPageViewModel::class.java)
                     mutableStateOf(viewModel)
                 }
+                onInitialize(devicesPageViewModel)
 
-                DevicesPage(
-                    windowScope = windowScope,
+                DevicesPage(windowScope = windowScope,
                     devicesPageViewModel = devicesPageViewModel,
                     onNavigateSetting = { mainViewModel.selectPage(Page.SettingPage) },
-                    onNavigateDevice = { mainViewModel.selectPage(Page.DevicePage(it)) }
-                )
+                    onNavigateDevice = { mainViewModel.selectPage(Page.DevicePage(it)) })
             }
             Page.SettingPage -> {
                 val settingPageViewModel by remember {
@@ -84,12 +92,10 @@ private fun MainPages(windowScope: WindowScope, mainViewModel: MainContentViewMo
                 }
                 onInitialize(settingPageViewModel)
 
-                SettingPage(
-                    windowScope = windowScope,
+                SettingPage(windowScope = windowScope,
                     settingPageViewModel = settingPageViewModel,
                     onNavigateDevices = { mainViewModel.selectPage(Page.DevicesPage) },
-                    onSaved = { mainViewModel.refreshSetting() }
-                )
+                    onSaved = { mainViewModel.refreshSetting() })
             }
             is Page.DevicePage -> {
                 val devicePageViewModel by remember {
@@ -100,11 +106,9 @@ private fun MainPages(windowScope: WindowScope, mainViewModel: MainContentViewMo
                 }
                 onInitialize(devicePageViewModel)
 
-                DevicePage(
-                    windowScope = windowScope,
+                DevicePage(windowScope = windowScope,
                     deviceViewModel = devicePageViewModel,
-                    onNavigateDevices = { mainViewModel.selectPage(Page.DevicesPage) }
-                )
+                    onNavigateDevices = { mainViewModel.selectPage(Page.DevicesPage) })
             }
         }
     }
@@ -156,5 +160,29 @@ private fun Message.toStringMessage(): String {
         is Message.StartRecordingMovie -> "Start recording movie on ${this.context.displayName}"
         is Message.StopRecordingMovie -> "Stop recording movie on ${this.context.displayName}"
         is Message.FailedRecordingMovie -> "Failed recording movie on ${this.context.displayName}"
+    }
+}
+
+@Composable
+private fun onInitialize(viewModel: ViewModel) {
+    DisposableEffect(viewModel) {
+        viewModel.onStarted()
+        onDispose {
+            viewModel.onCleared()
+        }
+    }
+}
+
+@Composable
+fun AppWindow(onCloseRequest: () -> Unit, state: WindowState, content: @Composable FrameWindowScope.() -> Unit) {
+    Window(
+        onCloseRequest = onCloseRequest,
+        state = state,
+        resizable = false,
+        undecorated = true,
+        transparent = true,
+        icon = painterResource(Images.DEVICE),
+    ) {
+        Card(shape = RoundedCornerShape(8.dp)) { content.invoke(this) }
     }
 }
