+++ src/main/kotlin/view/components/molecules/TitleAndRadioButtons.kt
@@ -2,11 +2,9 @@ package view.components.molecules
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -22,13 +20,11 @@ fun TitleAndRadioButtons(
     onSelect: (String) -> Unit,
     modifier: Modifier = Modifier
 ) {
-    Box(modifier = modifier) {
-        Column(modifier = Modifier.padding(8.dp), verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            Texts.Subtitle1(subtitle1)
-            Texts.Subtitle2(subtitle2)
-            Spacer(modifier = Modifier.height(4.dp))
-            RadioButtons(selectedItem = selectedItem, items = items, onSelect = { onSelect(it) })
-        }
+    Column(modifier = modifier, verticalArrangement = Arrangement.spacedBy(8.dp)) {
+        Texts.Subtitle1(subtitle1)
+        Texts.Subtitle2(subtitle2)
+        Spacer(modifier = Modifier.height(4.dp))
+        RadioButtons(selectedItem = selectedItem, items = items, onSelect = { onSelect(it) })
     }
 }
