+++ src/main/kotlin/viewmodel/MainContentViewModel.kt
@@ -17,7 +17,7 @@ import model.usecase.GetSystemDarkModeFlowUseCase
 import model.usecase.IsSetupCompletedUseCase
 import resource.Strings.NOT_FOUND_ADB_COMMAND
 import resource.Strings.NOT_FOUND_SCRCPY_COMMAND
-import view.page.Page
+import view.components.pages.Page
 
 class MainContentViewModel(
     private val fetchSettingUseCase: FetchSettingUseCase,
